#82 moved new classes in separate package
