+++ src/main/scala/dbtarzan/gui/table/headings/TableColumnsIWithIndex.scala
@@ -1,4 +1,4 @@
-package dbtarzan.gui.table
+package dbtarzan.gui.table.headings
 
 import scala.collection.mutable
